ai-agents/v2/domain-specific-ai-agents/layerzero/requirements.txt
python-dotenv
langchain
langchain_openai
pypdf
faiss-cpu
uvicorn
langchain_community
fastapi